Add new tests to the ChecksCatalog mocking axious
