+++ assets/package.json
@@ -14,6 +14,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "autoprefixer": "^10.4.13",
+    "axios-mock-adapter": "^1.21.2",
     "babel-jest": "^29.2.1",
     "babel-loader": "^9.1.0",
     "esbuild": "^0.15.15",
